Fixed a problem with missing </a> and parsec error
Ran all the code through here: https://validator.w3.org/

After validating the HTML I fixed a parsec error (which honestly wasn't an error because it was just a HTML comment, but for some reason it came out as an error so I just moved out the comment to stop seeing it come up). And the <a> tag before the button wasn't closed and raising errors. So I closed it. Thank god I now know how to validate HTML. Yay